Migrate tools test fakes to null safety (#80154)
