+++ packages/flutter_tools/test/general.shard/bundle_shim_test.dart
@@ -14,7 +14,7 @@ import 'package:flutter_tools/src/globals_null_migrated.dart' as globals;
 
 import '../src/common.dart';
 import '../src/context.dart';
-import '../src/fakes.dart';
+import '../src/test_build_system.dart';
 
 // Tests for the temporary flutter assemble/bundle shim.
 void main() {
