+++ packages/flutter_tools/test/general.shard/dart/generate_synthetic_packages_test.dart
@@ -16,7 +16,7 @@ import 'package:flutter_tools/src/build_system/targets/localizations.dart';
 
 import '../../src/common.dart';
 import '../../src/fake_process_manager.dart';
-import '../../src/fakes.dart';
+import '../../src/test_build_system.dart';
 
 void main() {
   testWithoutContext('calls buildSystem.build with blank l10n.yaml file', () async {
