+++ packages/flutter_tools/test/general.shard/ios/ios_device_start_prebuilt_test.dart
@@ -23,6 +23,7 @@ import 'package:flutter_tools/src/ios/mac.dart';
 import 'package:mockito/mockito.dart';
 
 import '../../src/common.dart';
+import '../../src/fake_devices.dart';
 import '../../src/fake_process_manager.dart';
 import '../../src/fakes.dart';
