+++ packages/flutter_tools/test/general.shard/protocol_discovery_test.dart
@@ -11,7 +11,7 @@ import 'package:fake_async/fake_async.dart';
 
 import '../src/common.dart';
 import '../src/context.dart';
-import '../src/fakes.dart';
+import '../src/fake_devices.dart';
 
 void main() {
   group('service_protocol discovery', () {
