+++ packages/flutter_tools/test/general.shard/resident_web_runner_cold_test.dart
@@ -26,6 +26,7 @@ import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 import '../src/common.dart';
 import '../src/context.dart';
 import '../src/fakes.dart';
+import '../src/test_build_system.dart';
 
 void main() {
   ResidentWebRunner residentWebRunner;
